+++ plugins/postcss-base-plugin/src/index.ts
@@ -55,3 +55,4 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };