+++ plugins/postcss-color-function-display-p3-linear/src/index.ts
@@ -105,3 +105,4 @@ const postcssPlugin: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 postcssPlugin.postcss = true;
 
 export default postcssPlugin;
+export { postcssPlugin as 'module.exports' };