+++ plugins/postcss-logical/src/index.ts
@@ -190,3 +190,4 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };