+++ plugins/postcss-minify/src/index.ts
@@ -200,3 +200,4 @@ const creator: PluginCreator<pluginOptions> = () => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };