+++ plugins/postcss-position-area-property/src/index.ts
@@ -24,3 +24,4 @@ const creator: PluginCreator<pluginOptions> = () => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };