+++ plugins/postcss-random-function/src/index.ts
@@ -48,3 +48,4 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 creator.postcss = true;
 
 export default creator;
+export { creator as 'module.exports' };