+++ plugins/postcss-light-dark-function/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to PostCSS Light Dark Function
 
+### Unreleased (major)
+
+- Updated: Support for Node `20.19.0` or later (major).
+- Removed: `commonjs` API. In supported Node versions `require(esm)` will work without needing to make code changes.
+
 ### 2.0.11
 
 _September 21, 2025_